+++ bevy_rapier2d/examples/joints_despawn2.rs
@@ -24,7 +24,7 @@ fn main() {
         .run();
 }
 
-fn setup_graphics(mut commands: Commands, mut configuration: ResMut<RapierConfiguration>) {
+fn setup_graphics(mut commands: Commands) {
     let mut camera = OrthographicCameraBundle::new_2d();
     camera.transform = Transform {
         translation: Vec3::new(0.0, -200.0, 0.0),
