+++ bevy_rapier2d/examples/multiple_colliders2.rs
@@ -44,13 +44,11 @@ pub fn setup_physics(mut commands: Commands) {
     let centery = shift / 2.0;
 
     let mut offset = -(num as f32) * (rad * 2.0 + rad) * 0.5;
-    let mut color = 0;
 
     for j in 0usize..20 {
         for i in 0..num {
             let x = i as f32 * shift * 5.0 - centerx + offset;
             let y = j as f32 * (shift * 5.0) + centery + 3.0;
-            color += 1;
 
             commands
                 .spawn()
