+++ bevy_rapier3d/examples/ray_casting3.rs
@@ -56,7 +56,6 @@ pub fn setup_physics(mut commands: Commands) {
     let centerz = shift * (num / 2) as f32;
 
     let mut offset = -(num as f32) * (rad * 2.0 + rad) * 0.5;
-    let mut color = 0;
 
     for j in 0usize..20 {
         for i in 0..num {
@@ -64,7 +63,6 @@ pub fn setup_physics(mut commands: Commands) {
                 let x = i as f32 * shift - centerx + offset;
                 let y = j as f32 * shift + centery + 3.0;
                 let z = k as f32 * shift - centerz + offset;
-                color += 1;
 
                 // Build the rigid body.
                 commands
